feat: improved labeling for very curved lines
feat: improved labeling for very curved lines

Changed the `make_labels.py` file to check if the sliding windows (after the first five windows) were against the sides of the image. The previous model always fails on very curved lines as once the side of the image is hit, the other windows continue straight up in the image. This version also includes a higher amount of windows (35 vs. 15 previously) and adds the second search around the initial line that can help further with curves. 

Known issue: This does sometimes create issues on straight lines if there are large distance between lines.