feat: fully convolutional NN
Added a file containing a fully convolutional neural network that, once optimized, may be used to directly show the car's lane on a given road image. This solves the issue caused by the specific perspective transformation needed on a given image, as this will potentially draw the lane back onto the exact space for the given image.